suggestion_box/requirements.txt
django
django-dotenv